Removed test modul as it may break causing false negatives. Set default trust_remote_code to false. Added an example that highlights the usecase.
